refactor: 数据统计模块优化调整 (#161)

+++ types/product_accessorie.go
@@ -16,7 +16,7 @@ type ProductAccessorieWhere struct {
 	Type       enums.ProductAccessorieType       `json:"type" label:"类型" find:"true" create:"true" update:"false" info:"true" sort:"6" type:"number" input:"select" required:"true" preset:"typeMap"`          // 类型
 	RetailType enums.ProductAccessorieRetailType `json:"retail_type" label:"零售方式" find:"true" create:"true" update:"false" info:"true" sort:"7" type:"number" input:"select" required:"true" preset:"typeMap"` // 零售类型
 	Remark     string                            `json:"remark" label:"备注" find:"true" create:"true" update:"true" info:"true" sort:"8" type:"string" input:"textarea" required:"false"`                       // 备注
-	Price      decimal.Decimal                   `json:"price" label:"单价" find:"false" create:"true" update:"true" info:"true" sort:"9" type:"decimal" input:"number" required:"true"`                         // 单价
+	Price      decimal.Decimal                   `json:"price" label:"单价" find:"false" create:"true" update:"true" info:"true" sort:"9" type:"float" input:"number" required:"true"`                           // 单价
 	Status     enums.ProductAccessorieStatus     `json:"status" label:"状态" find:"true" create:"false" update:"false" info:"true" sort:"10" type:"number" input:"select" required:"true" preset:"typeMap"`      // 状态
 
 	EnterId string `json:"enter_id" label:"入库单" find:"false" create:"false" update:"false" sort:"11" type:"string" input:"text" required:"true"` // 入库单
