+++ types/product_accessorie_enter.go
@@ -66,7 +66,7 @@ type ProductAccessorieEnterReqProduct struct {
 	Name       string                            `json:"name" label:"名称" find:"true" create:"true" update:"false" sort:"3" type:"string" input:"text" required:"true" binding:"required"`                             // 名称
 	Type       enums.ProductAccessorieType       `json:"type" label:"类型" find:"true" create:"true" update:"false" sort:"4" type:"number" input:"select" required:"true" preset:"typeMap" binding:"required"`          // 类型
 	RetailType enums.ProductAccessorieRetailType `json:"retail_type" label:"零售方式" find:"true" create:"true" update:"false" sort:"5" type:"number" input:"select" required:"true" preset:"typeMap" binding:"required"` // 零售类型
-	Price      decimal.Decimal                   `json:"price" label:"单价" find:"false" create:"true" update:"true" sort:"6" type:"decimal" input:"number" required:"false"`                                           // 单价
+	Price      decimal.Decimal                   `json:"price" label:"单价" find:"false" create:"true" update:"true" sort:"6" type:"float" input:"number" required:"false"`                                             // 单价
 	Remark     string                            `json:"remark" label:"备注" find:"true" create:"true" update:"true" sort:"6" type:"string" input:"textarea" required:"false"`                                          // 备注
 	Stock      int64                             `json:"stock" label:"库存" find:"false" create:"true" update:"true" sort:"7" type:"int" input:"number" required:"true" binding:"required"`                             // 库存
 }
